multimodal-assistant/requirements.txt
transformers
torch
Pillow
PyPDF2
python-docx
faiss-cpu
sentence-transformers
